make file size long and refactor how paramateres are calculated

+++ src/main/java/terasort/hawk/iit/edu/FinalChunkFileWriter.java
@@ -8,10 +8,10 @@ public class FinalChunkFileWriter extends Thread {
 	private ChunkBuffer chunkBuffer;
 	private String filename;
 	private Integer chunkSize;
-	private Integer fileSize;
+	private Long fileSize;
 
 	public FinalChunkFileWriter(ChunkBuffer chunkBuffer, String threadName, String filename, Integer chunkSize,
-			Integer fileSize) {
+			Long fileSize) {
 		this.chunkBuffer = chunkBuffer;
 		this.setName(threadName);
 		this.filename = filename;
